Refactor Login auth system
